requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
Flask>=2.0.0
Werkzeug>=2.0.0 